day36/day36-connect-monitoring/requirements.txt
flask==3.0.3
prometheus-client==0.20.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
pyyaml==6.0.1
pandas==2.2.2
numpy==1.26.4
psutil==5.9.8
pytest==8.2.1
pytest-asyncio==0.23.7
